[#166] refactor: 인챈트 정보 폴더 구조 변경
