+++ src/app/_features/market/components/EnchantDropAndDetail.tsx
@@ -1,17 +1,17 @@
 'use client';
+import EnchantInfo from '@/app/_components/iteminfo/enchant/EnchantInfo';
 import {
   EnchantStoreType,
   useSelectEnchantStore,
 } from '@/app/_store/selectEnchantStore';
-import React from 'react';
 import EnchantDropList from './EnchantDropList';
-import EnchantInfo from '../iteminfo/enchant/EnchantInfo';
+import Column from '@/app/_components/layout/Column';
 
 const EnchantDropAndDetail = () => {
   const enchant = useSelectEnchantStore((state) => state.enchant);
 
   return (
-    <>
+    <Column className="w-full max-w-[400px] gap-2">
       {enchant && (
         <>
           <EnchantDropList enchantData={enchant as EnchantStoreType} />
@@ -20,7 +20,7 @@ const EnchantDropAndDetail = () => {
           </div>
         </>
       )}
-    </>
+    </Column>
   );
 };
